Rename runlint.py to allow for other linters to be added here as well.
Summary:
We will add a few python linters, as well as a 'master' linter that
can handle any language.

I also cleaned up lint errors while in the area, so the script is a
bit rewritten.

Also, unrelatedly, updated requirements.txt to add some dependencies
we will be needing shortly.

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D172