Resolve a TODO by adding an explicit new mode 'auto', for --blacklist.
Summary:
In auto mode, we use the blacklist when expanding directories, but not
when linting files listed on the commandline.  This is the default.

Add a new flag, --blacklist-filename, to override the default blacklist.

Reviewers: benkomalo, chris

Differential Revision: http://phabricator.khanacademy.org/D174

Note: I also fixed up a merge as part of this commit.  I don't know
exactly how I got this wrong, I think when trying to resolve merge
conflicts somewhere, but the entire contents of
40fd9444d7c09600474f311d675f3f6950cb1a05 were never properly applied
(or something)?  'git show 40fd9444' will show, for instance, that the
helpstring for --blacklist refers to 'auto', but 'auto' was not added
to 'choices'.  I don't know how that happened, but I made the changes
to make this commit an exact duplicate of
6cc22ae733a88679881ac5aa0e369b8d44cfb0b1, which I was trying to
cherry-pick in from another branche.  This is kind of a manual
cherry-pick.

Tested: git show 6cc22ae733a88679881ac5aa0e369b8d44cfb0b1:runlint.py | diff -u - runlint.py