Fix a typo involving overload of the 'lang' var, and add 'verbose' mode.
The linter had stopped linting!  Eventually I figured out it was
because I had renamed options.lang to lang, which conflicted with an
existing 'lang' in the function.  I renamed one of them to fix that.

In the process of debugging, I found some verbose output was helpful.
I left it in under a new --verbose flag.