Modify hghook to call runlint instead of gjslint, so it can lint python too.
Summary:
I also modified hghook substantially to fit the current python style
guide.  It now passes lint itself. :-)

I augmented the README with the new control flow and other new
information.

I modified runlint.py:main() to make it easier to call it from other
python files.  Before, it required an options object like the kind
produced by optparse, which wasn't necessarily that practical for
other scripts.  I separated it out to its constituent flags.

Change runlint to silently skip files it doesn't know how to lint.

Before, it would print a message saying 'skipping this file I don't
know how to lint.'  But when run from phabricator or hg, which just
pass in all the files in a commit, we're likely to run into this
'unknown language' situation a lot, and the right thing to do is to
silently ignore those files, not to print verbiage.

Fix a typo involving overload of the 'lang' var, and add 'verbose' mode.

The linter had stopped linting!  Eventually I figured out it was
because I had renamed options.lang to lang, which conflicted with an
existing 'lang' in the function.  I renamed one of them to fix that.

In the process of debugging, I found some verbose output was helpful.
I left it in under a new --verbose flag.

Add a few more files that need to be blacklisted.

We need to get this live soon so we don't have to keep updating the
blacklist! :-}

Fix some typos/etc as found by benkomalo in review.

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D186