Change the way blacklists are treated to allow them to be in-repository.
Summary:
Before this change, the blacklist file had to live in khan-linter, or
be explicitly specified on the commandline.

Now, the blacklist file can have the text '<ancestor>/' in it, in
which case it is taken to be in the same directory as the file being
linted, or some parent directory of it.  This allows us to have
different blacklists for different files, and even to have multiple
blacklists in use for one lint run!  (For instance, the 'main' repo
might have one blacklist, but each of the sub-repos/sub-modules might
have a blacklist of their own.)

The default blacklist has '<ancestor>/' in it, so this is the new
scheme from now on.  I've already committed a blacklist into the
stable repo so we can start using this scheme right away.

I totally messed up my rename of runlint_blacklist -> lint_blacklist before.

Don't ask me how.  I'm clearly still a git newbie.

I added some text at hte top of lint_blacklist.txt indicating that
this is an example file.

Test Plan:
manually tested on directories with a blacklist file in the
same directory as the file being tested, in a parent directory, and
not existing at all.  Tested when there are two blacklist files in a/
and a/b, and made sure both a/foo.py and b/foo.py picked up the right
blacklist file.

Reviewers: chris, benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D194