Fix the linting of directories when --blacklist=yes (and maybe auto).
Summary:
Before, we added directories to the directories_to_lint regardless of
their blacklist status.  This means we'd lint all files in the
directory even if the blacklist said not to.

Now, we treat directories in the same control-flow we use for files,
so we handle the blacklist properly.

Test Plan:
~/khan/khan-linter/runlint.py --verbose --blacklist=no third_party
and also with --blacklist=yes and --blacklist=auto

Reviewers: dylan

Reviewed By: dylan

Differential Revision: http://phabricator.khanacademy.org/D637