Add a new exception for long json lines in docstrings.
Summary:
The new documenting-apis-in-docstrings effort I'm starting requires us
to put example json output in our docstrings.  json output is often
longer than 80 characters.  To make this pass lint, I add an exception
to our lint-checker.  It does a reasonable, though far from perfect,
job of determining if a long line is both (a) json, and (b) inside a
docstring.

Fix the mini-syntax-checker to allow for multi-lines definitions.

Before, we needed it to look like
   def foo(...):
      """...
      longline

Now it can look like
   def foo(...,
           more...):
      """...
      longline

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D736