Add linting of html files.
Summary:
For now, the only linting that is done is checking for the |static_url
modifier on static content.

The static_url filter, in url_util.py, is designed for any 'static'
content on our site; that is, content that is served via a static
handler in app.yaml, or otherwise is known to not either need or set
cookies.  Cookies don't do any good for static content, so there's no
reason to send them as part of the request, or as part of the
response; the static_url filter makes this happen (by serving the
images from a separate-but-equal domain name that doesn't have cookies
on it).  Since cookies can be rather large, this speeds up the serving
of images.

The KA code that embeds <script> and <style> tags in our emitted html
is written to automatically apply static_url() when putting filenames
in those tags, so .js and .css files already take advantage of this
functionality.  But other static content, particularly images, tend to
be specified directly in our html templates, where it's harder for the
code to automatically fix them up.  Instead, we use a lint check to
ensure that static content in .html files are marked with the
static-url filter, like so:
   <img width="100%" src="{{"/images/toolkit/summary.jpg"|static_url}}" />
   .container { background-image:url({{"/images/indicator.png"|static_url}}); }

Simplify the code by using re.finditer().

I hadn't known about finditer() before.  Nice!

Get rid of unused 'colnum' var.

Add a comment to explain some tricky-ish code.

Add a TODO

Test Plan:
Ran
   ./static_content_refs.py /home/csilvers/khan/stable/templates/
and verified all the errors looked reasonable.  Then ran
   ./runlint.py
and verified it gave the same set of errors.  Then ran
   ./static_content_refs.py --fix /home/csilvers/khan/stable/templates/
and ran 'hg diff' and verified it correctly fixed all the errors
identified.

I was conservative in my regexp matching, so there may be more places
we could add |static_url that this misses.

Reviewers: chris, benkomalo, kamens

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D966