Add in a check for leftover debug code.
This puts in a crude check for the token "console.log" or
"window.console.log", and errors out if so. It does not update the
autofixer to remove those references.

Test Plan:
Ran js_runlint.py against a file with console.log("foo"); in it (error)
Ran js_runlint.py against a file with window.console.log("foo"); in it (error)
Ran js_runlint.py against a file with this.console.log("foo"); in it (no error)