Suppress lint checks on merge.
Summary: Look for MERGE_HEAD explicitly rather than implicitly via the commit message.

Test Plan:
Created a sub-branch of this one, and in that sub-branch edited
githook.py in a way that conflicts with this change, merged this
change into that, resolved the merge error, and did a git commit, and
saw it say 'Skipping lint on merge'.

Created a sub-branch of this one, and in that sub-branch edited
githook.py in a way that conflicts with this change, merged this
change into that, resolved the merge error, and did a git commit, and
saw it say 'Skipping lint on merge'.

Reviewers: benkomalo, alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D2259