Remove comment lines when saving to .git/commit.save.
Summary:
'git commit -F' does not remove comment lines, and the most likely way
people will use .git/commit.save is with -F, so we need to remove the
comment lines for it.

Test Plan:
Put in an extra blank line, which should cause this commit to fail
with a lint error.  Then I removed the blank line and re-committed
with -F.  If there are no comment-lines in the resulting commit,
everything worked!

Reviewers: benkomalo, alpert

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D2321