Add the ability of khan-linter to run a second linter script.
Summary:
We will use this to have it run tools/runlint.py when in the webapp
tree.  This will let us run runlint.py with git/hg commits, and
phabricator diffs, in addition to 'make lint' like we do now.

However, I've commented out the functionality for now because
tools/runlint.py is rather slow, due to the lint-file discovery
technique it uses (basically, searching under all of webapp).  Also,
exercises.babel_lint.lint_json_output_from_all_files ignores its
arguments and is thus always slow.  Will have to fix those problems
before enabling this.

Rename _blacklist_filename now that it is used for the extra-linter as well.

We now call it _resolve_ancestor, since the salient point is that it
deals with '<ancestor>' strings.

Test Plan:
Ran
   % ~/khan/devtools/khan-linter/runlint.py -v api/v1_test.py --extra-linter='<ancestor>/tools/runlint.py' ~/khan/devtools/khan-linter/runlint.py
   Using blacklist <ancestor>/lint_blacklist.txt for files under directories
   Considering /home/csilvers/khan/webapp/api/v1_test.py: blacklist None ... LINTING
   Considering /home/csilvers/khan/devtools/khan-linter/runlint.py: blacklist None ... LINTING
   --- linting /home/csilvers/khan/devtools/khan-linter/runlint.py (python)
   --- linting /home/csilvers/khan/webapp/api/v1_test.py (python)
   --- running extra linter: /home/csilvers/khan/webapp/tools/runlint.py /home/csilvers/khan/webapp/api/v1_test.py
   Running i18n_lint.lint_non_literal_i18n ... 0 errors
   Running i18n_lint.lint_non_literal_i18n_do_not_translate ... 0 errors
   Running i18n_lint.lint_templates_are_translated ... 0 errors
   Running jinja2_lint.lint_no_jinja2_markup_in_python ... 0 errors
   Running less_lint.lint_only_use_import_once ... 0 errors
   Running jinja2_lint.lint_no_templatetags_calls_in_python ... 0 errors
   Running exercises.babel_lint.lint_json_output_from_all_files ... 0 errors
   Running i18n_lint.lint_missing_safe_in_jinja2 ... 0 errors

The important thing to note is that it ran runlint.py, but only for
v1_test.py; it didn't run it for khan-linter's runlint.py, because
there's no tools/runlint.py in the khan-linter repo.

Ran
   make lint
in webapp with no failures, indicating the blacklist is still being
parsed properly.

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D2840