Switch from closure linter to jshint
Summary:
Google's closure linter isn't as flexible as jshint. It doesn't catch a bunch of things that jshint does (in particular, I'm looking forward to jshint's support for catching undeclared globals) (and isn't as popular either, to boot).

This isn't quite ready to land yet because it's slower than I'd like and I haven't fixed up all or JS to conform yet.

Reviewers: benkomalo, joel, csilvers

Reviewed By: csilvers

CC: csilvers

Differential Revision: http://phabricator.khanacademy.org/D4028