Don't require a 'Test plan:' section if the commit message is 'WIP'.
Summary:
For the phabricator workflow, some people always have the git commit
message be 'WIP', and put in the actual message at 'arc diff' time.
We don't require a 'real' commit message in that case.

Oops, fix a typo that snuck in after testing.

Test Plan:
Originally committed this message with
   git commit -am 'test'
and it failed with 'missing test plan'.  Then did
   git commit -am 'wip'
and it succeeded.
(Then I ran git commit --amend to put in this real message :-) )

Ran
   git commit -am 'true'
and saw it fail.  Then ran
   git commit -am 'wip do'
and saw it succeed.

Reviewers: benkomalo

Reviewed By: benkomalo

CC: pamela

Differential Revision: http://phabricator.khanacademy.org/D4469