Give correct lint output when jsx encounters a parse error.
Summary:
Sometimes when jsx can't compile a file, it's due to some internal
error or something.  That was the assumption before, when we always
raised a RuntimeError in that case.

However, sometimes jsx can't compile a file due to a parse error.  If
that happens, the proper thing to do is to report that as a lint
error, not raise an exception.  Now we look for that case and properly
report it.

Test Plan:
Changed editor-status-view.jsx to have the line:
                        label=$._("Share link...")
(instead of label={$._("Share link...")}) and then ran
   ~/khan/devtools/khan-linter/runlint.py /home/csilvers/khan/webapp/javascript/editor-package/editor-status-view.jsx

After this change, it now reports:
   javascript/editor-package/editor-status-view.jsx:307: E=parse= XJS value should be either an expression or a quoted XJS text
and has a return-code of 1, both of which are as desired.

Reviewers: joel, tom

Reviewed By: joel

Differential Revision: http://phabricator.khanacademy.org/D4503