Run extra-linter once, not lots of times.
Before, we were calling
   <extra_linter> <file-to-lint> <file-to-lint> ...
and due to limits on the size of argv, this means we'd have to run
extra_linter dozens of times when there are lots of files to lint
(which there always are).

Now, we call
   <extra_linter> -
and pass in the files-to-lint on stdin.  All the extra-linters we use
-- ok, there's only one -- have already been rewritten to support this
new syntax.  The end result is that we have to run extra-linter only
once, rather than lots of times.

This matters for khan webapp, since the startup time of the webapp
extra-linter is non-trivial (it needs to walk the entire webapp file
tree looking for *_lint.py files).  The time for 'make lint' in webapp
has dropped from about 2 minutes to 37 seconds.

Auditors: chris

Test Plan:
In webapp-root, ran
    ~/khan/devtools/khan-linter/runlint.py -v
and saw it complete successfully, along with the log message saying it
was running the extra-linter.