Do some better short-cutting of the git lint check.
Summary:
In particular, if the user tries to commit an empty commit message, or
leaves it unchanged from the template, abort right away.

This commonly happens to me when I do 'git commit -a', and see from
the comments in the editor I'm accidentally committing a file I didn't
mean to.  I exit the editor to abort the commit, but first the githook
does a bunch of lint checks so it can take awhile.  Now it will abort
fast.

Fix comments based on review feedback.

Test Plan:
For this commit, I did
   git commit -a
and removed the text of the file up until the comment lines, and
exited the editor, and got the 'empty commit' message.

Then I did
   git commit -a
and exited the editor right away, and got the 'unchanged commit'
message.

Then I did
   git commit -a
and got rid of the 'test plan' line and wrote 'test' as the first
line, and exited the editor and got the 'Test plan: needed' message,
showing that stuff still works correctly.

(same)

Reviewers: benkomalo, alpert

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D5380