Prevent two linters from trying to update runlint.py at the same time.
Summary:
We have this auto-update feature which does a git pull.  This is nice,
but Phabricator often runs two linters at the same time, meaning they
could both contend for the 'git pull' call.  And in fact, we were
seeing problems with people reporting a 'git is locked' error when
running the linter.

This file-lock should resolve that.  I can't reproduce the problem
locally, so I'm not positive it will solve things, but it *should*,
dammit.

Test Plan: Fingers crossed.

Reviewers: alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D6529