Add a lint check to prevent committing a file with a merge conflict marker.
Summary:
If you do a merge and there are conflicts, git adds text like:

    <<<<<<< HEAD:file.txt
    Hello world
    =======
    Goodbye
    >>>>>>> 77976da35a11db4580b80ae27e8d65caf5208086:file.txt

You're supposed to resolve these before committing, but it's easy to
miss one, especially if a file has many merge sections.  git really
ought to catch that, but it doesn't, so we do in our lint check.

Test Plan:
I manually added this text to the middle of runlint.py:

    """
    <<<<<<<
    >>>>>>> HEAD
    """

and ran
   ~/khan/devtools/khan-linter/runlint.py ~/khan/devtools/khan-linter/runlint.py
and saw it say:
   ../devtools/khan-linter/runlint.py:321:1: E1 git conflict marker "<<<<<<<" found
   ../devtools/khan-linter/runlint.py:322:1: E1 git conflict marker ">>>>>>>" found

and I checked and indeed those two lines were 321 and 322.

Reviewers: chris, alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D7409