Turn on linting of unused variables.
Summary:
The "vars" value for "unused" complains about unused variables, but not about
unused function arguments.

One case where it's weird to complain about unused function arguments is when we
have RequireJS wrappers with the sugar syntax, and require() isn't called within
the function:

    define(function(require) {
        // Nothing calls require in here
    });

So let's just get rid of unused variables for now, since that tends to be the
more confusing thing in code, and occasionally function variables are left in
for callbacks for readability purposes.

Test Plan: N/A

Reviewers: alpert, csilvers

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D13552