Recommend 'git recommit' to re-use a commit message.
Summary:
Sometimes people complain that they craft a beautiful commit message,
only to have it rejected due to lint errors.  The beautiful commit
message isn't gone though!  You can easily re-use it via 'git
recommit'.  I advertise that fact now.

The alternative would be to split this check in two: the lint check
that runs before the commit message editor pops up, and the message
check that runs after.  I dislike this approach though because if
linting takes a long time, there's a long delay between when 'git
commit' is typed and the editor pops up.  It's easy to switch off to
another task and then forget to come back to the commit.

Better, I think, to type the commit message right away, so the
'potentially long time' takes place after the user has done everything
they need to do.

Test Plan:
Ran
   git commit -am test
on this commit and had it say:

   Missing "Test plan:" or "Review:" section in the commit message.

   --- 1 lint errors ---
   Commit message saved to .git/commit.save
   Use "git recommit -a" when the errors are fixed, to re-use this commit message

Reviewers: chris

Subscribers: #infrastructure

Differential Revision: https://phabricator.khanacademy.org/D17655