vendor all dependencies
Summary:
This vendors all dependencies directly into the repository (Python and
Node), so that there is zero setup and the application will work no
matter what the user current virtualenv environment setup is, etc.

For Node, its a simple matter of checking in the node_modules directory.
For python, there are some minor import path and makefile modifications.

Note to code reviewers, when reviewing, you can ignore the contents of
the /node_modules and /vendor directories, as they are just the vendored
libs.

This puts a tiny bit more work on the maintainers of this application to
commit any future generated dep changes along with changes to the
requirements files, but should make things significantly more robust for
end users. As long as they having Python and Node installed everything
should be kosher (we could vendor those too, but then we'd need to make
platform-specific distributions.)

As such, the post-pull update hook to re-run dependency installation
canhaz been removed.

Once this is accepted, we will need to modify khan-dotfiles to no longer
run requirements installation for this tool.

NOTE: phabricator apparently pukes on diffs with too many files, so the
actual vendored directories are left out of this commit, but I will squash it
back in after review I guess.

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: ethan, alpert, marcos

Differential Revision: https://phabricator.khanacademy.org/D18066