make khan-linter usable from default path
Summary:
Currently khan-linter works by the path to the runlint.py being
hardcoded anywhere that needs to use it.  We should allow it to be added
to the user default $PATH so that they can a.) easily call it manually
on a file, b.) reference it in hooks, scripts, etc without it being
brittle in a way that relies on it living in a specific location on
every system.

This change really just adds a simple bin/ directory with a single
symlink "ka-lint" to runlint.py, which is a descriptive name since you
won't have the parent directory context.  Other than that there are a
few minor changes to make the code symlink aware.

Related to Asana task:
https://app.asana.com/0/31769942937408/34648822879760

Test Plan:
- verify hardcoded paths to runlint.py still function
- add $PROJECT/bin to system $PATH, and verify "ka-lint" command
  functions as expected.
- run across entire webapp repo, verify output is identical to previous version.

Reviewers: benkomalo, csilvers, marcos

Reviewed By: csilvers, marcos

Subscribers: marcos, ethan

Differential Revision: https://phabricator.khanacademy.org/D18095