Add auto-fixing for the jsx closing bracket location.
Summary:
We can do this now that Alex changed the linter to report enough
information to say where the closing bracket should go!

We handle both the case the bracket should move on the same line, and
that it should go on the next line.

Test Plan:
In webapp, I edited
   javascript/typeahead-package/search-results-popup.jsx
to remove `react/jsx-closing-bracket-location` from the
`eslint-disable` comment at the top of the file.  Then I ran
   arc lint javascript/typeahead-package/search-results-popup.jsx
and said 'yes' to take the suggestions, and did `git diff` to see
that everything was now indented properly.

Reviewers: alex

Reviewed By: alex

Differential Revision: https://phabricator.khanacademy.org/D23642