Suppress warnings when we .eslintignore a file.
When eslint skips over a file because it's on its ignore-list, it
prints a warning saying it's doing so.  This causes ka-lint to fail.
We now suppress that warning: we know what we're doing with our
ignore-list!

Auditors: jlfwong

Test Plan:
Ran this without error:
   cd khan-exercises && ka-lint

Before, it would complain a lot about skipping files in node_modules/.