Turn on the no-var rule!
Summary:
This is the rule that says to use const/let instead of var.

I previously added a bunch of eslint-disable directives to existing
code, so this new rule won't trigger on them.  I did this in webapp,
khan-exercises, and perseus.

Test Plan: Ran `ka-lint` in webapp with no errors.

Reviewers: jlfwong

Reviewed By: jlfwong

Subscribers: #infrastructure

Differential Revision: https://phabricator.khanacademy.org/D25652