Make extra linter search stop at repository boundary.
Summary:
We used to search for an extra `tools/runlint.py` file in an
ancestor directory and would run that as an extra lint check. This was a
problem for when I added the perseus submodule to webapp because it has
its own lint_blacklist.txt, so we do lint its files, but it doesn't have
its own tools/runlint.py so it would find webapp's tools/runlint.py,
which complained about a bunch of the files.

This makes the ancestor search for a tools/runlin.py stop when it
reaches a git or mercurial top-level directory. With this, we'll never
run the extra webapp linters except for files actually in webapp, and I
can commit my perseus submodule.

Test Plan:
 - Commit a change in webapp which updates the substate for the perseus
   submodule.
 - See that the commit-msg linter doesn't complain about i18n lint
   inside of perseus.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D27095