Update eslint to to fix object rest destructuring bug
Summary:
The existing version of eslint was marking object rest destructuring as
an error, e.g. with `const { prop1, prop2, ...rest } = obj;` it was
marking `rest` as being `undefined`.  Upgrading to 1.10.3 fixes this.
I tried upgrade to the 2.x line but it introduced a different issue,
namely it thought `...` was an invalid token.

Test Plan:
- ka-lint math-input/src/components/many-keypad-buttons.js (the offending file)
- make lint (in webapp to make sure it doesn't break anything else)

Reviewers: csilvers, charlie

Subscribers: