use babel-eslint in khan-linter
Summary:
This enables the linter to correctly parse class properties like `static propTypes = {...}`, enabled by the `es7.classProperties` Babel rule.
That's because it makes it much easier to define React classes in the new ES6 class syntax, which is the wave of the future! See some experimenting with it in D30730 (which is currently blocked on a lint failure that this change resolves).

That said, if this isn't a trivial decision, we don't need to worry about blocking D30730; as @csilvers notes, we could just Nolint that file for now.

NOTE: This diff doesn't include the corresponding changes to `node_modules`. Once the change is approved, I'll sneak the ~2.8k new files into the landing process.

Test Plan:
Load up D30730 in webapp. Before this change, `make linc` fails on a syntax error. After this change, `make linc` succeeds.
And then `make lint` the entire webapp, just to make sure we didn't break anything.

Reviewers: kevinb, csilvers, riley

Reviewed By: csilvers, riley

Subscribers: amy, csilvers

Differential Revision: https://phabricator.khanacademy.org/D30752